net: ip: Use rb trees in IPv4 reassembly and handle overlapping fragments.
IPv4 fragment reassembly is changed from using a linked list
to a rb tree. This helps faster inserting of fragments. If we
receive an overlapping fragment, we now discard the fragment
list. Overlapping fragments don't really have any usecase and
this simplifies the code and eliminates potential DDoS attack
vectors. Duplicate fragments are ignored.

Following are the commits that are taken from upstream.
1. https://source.codeaurora.org/quic/qsdk/oss/kernel/linux-msm/commit/?h=caf7/korg/stable/linux-4.4.y&id=ef0f963de1d2c5bc99d3d6ace3dd44a7d6002717
Discarding overlapping and duplicate fragments.

2. https://source.codeaurora.org/quic/qsdk/oss/kernel/linux-msm/commit/?h=caf7/korg/stable/linux-4.4.y&id=3f78a3f45e79ca378cb850a598e4c76633710e92
Implementation of rb trees in fragment reassembly storage.

3. https://source.codeaurora.org/quic/qsdk/oss/kernel/linux-msm/commit/?h=caf7/korg/stable/linux-4.4.y&id=acd00a0692072b374afee4b6f38c1eb1c6cf6f4a
Fixes a precedence bug in ip_expire() routine.

4. https://source.codeaurora.org/quic/qsdk/oss/kernel/linux-msm/commit/?h=caf7/korg/stable/linux-4.4.y&id=cb5fd4aa24b57206548d5940dc359f0b181a2688
Fixes a crash by setting a struct sk_buff {} variable “sk” to NULL.

5. https://source.codeaurora.org/quic/qsdk/oss/kernel/linux-msm/commit/?h=caf7/korg/stable/linux-4.4.y&id=873beab3740a085ba284ed7daa4ab4068074db53
Do not handle duplicate fragments as overlapping and ignore them.

6. https://source.codeaurora.org/quic/qsdk/oss/kernel/linux-msm/commit/?h=caf7/korg/stable/linux-4.4.y&id=9c6727de82e47238f389b415e25443ac46a06de5
This upstream change adds new member to netns_frag {}
structure. The code related to this change is not taken.
This is referenced since this interferes with the code and
local changes are made to go ahead with the old design.

7. https://source.codeaurora.org/quic/qsdk/oss/kernel/linux-msm/commit/?h=caf7/korg/stable/linux-4.4.y&id=826ff799146685450f84e3158ce66499c928c8ea
A member added to struct sk_buff in this upstream commit
is added to the code and an integration with the old design
has been followed.

Change-Id: Ief1e249894954a7a6e3f48d6c59e08c39fc98a5e
Signed-off-by: Gaurao Chaudhari <gaurchau@codeaurora.org>