net: cnss2: Do not free Target Dump data in wifi down path
Currently Target dump data is freed in wifi down path as part of
cnss_unregister_ramdump_v2. This causes an issue if a target assert
happens in down path as the dump data collected immediately after
target assert is freed subsequently because of which the target assert
cannot be debugged.

Do not free the target dump data in down path in cnss_unregister_ramdump_v2
and instead free it in cnss_register_ramdump_v2 before allocating again
in wifi up path.

Signed-off-by: Vignesh Viswanathan <viswanat@codeaurora.org>
Change-Id: If25eba2d43e4d21e8fe8123281613f56724dcc02