sound: soc: qcom: Fix ipq_config_mclk_sel API
This API was updating bit 1 in ADSS_STEREO0_STEREO0_CONFIG_REG register
after clearing bit 10.

For configuring mclk_sel we need to set bit 10 as 1 meaning it will be
external clk src. If we leave it as 0, mclk will be Audio PLL src.

This function should be called only when ipq is slave and mclk is
generated by master to be fed into ipq.

Signed-off-by: Selvam Sathappan Periakaruppan <speriaka@codeaurora.org>
Change-Id: Ib21747360e387935c1149e624be379fa96126952