net: cnss2: Trigger an assert if of_register_mhi_controller fails
Currently if of_register_mhi_controller fails in cnss_pci_register_mhi
which is called as part of cnss_pci_probe, we go to failure path in
cnss_pci_register_mhi where we try to disable PCI bus.
In cnss_pci_disable_bus, we try to access mhi_ctrl which would be NULL
because of_register_mhi_controller as failed. This leads to a NULL
pointer dereference.

To avoid this, if of_register_mhi_controller fails, trigger an assert.

Signed-off-by: Vignesh Viswanathan <viswanat@codeaurora.org>
Change-Id: If9378a41ac612e72910c182db92a2a9ac8fdb437