net: cnss2: Make recovery flow for QCN9000 same as AHB targets
Currently for QCN9000, when target assert notification is received
from MHI, target dump collection and upload is done followed by
subsystem_restart_dev sequence.

This leads to the wifi driver receiving SSR events after
the ramdump device has been destroyed. This is an issue if the
userspace relies on ACFG based notifications for dump collection as
the dump device would have already been destroyed when the ACFG dump
ready indication is sent.

Make the recovery flow of QCN9000 same as AHB targets with the below
changes
1. Send SUBSYS_PREPARE_FOR_FATAL_SHUTDOWN notification to wifi driver
   as soon as target assert notification is received from MHI if
   target recovery is enabled for that target.
2. In case target recovery is disabled, SUBSYS_PREPARE_FOR_FATAL_SHUTDOWN
   notification to wifi driver would be sent after RDDM is done and
   wifi driver would trigger a system crash.
3. Collect the RDDM segments in cnss_pci_collect_dump_info as part of
   cnss_do_recovery but do not create ramdump device. Instead start
   the subsystem_restart_dev sequence.
4. When cnss_subsystem_shutdown is called from SSR, skip shutdown if
   rddm data is collected. Shutdown will be done after dump upload.
5. As part of cnss_subsystem_ramdump, create the ramdump device and
   upload the dump file. After dump is complete, shutdown the target.
6. Add a cnss_get API to get the ramdump device name per target. This
   would be used by the wifi driver to notify the ramdump device name to
   the userspace.

Signed-off-by: Vignesh Viswanathan <viswanat@codeaurora.org>
Change-Id: Ic180edeab10b53611bf997fc98020a16ed4d6647