Revert "base: firmware-class: BUG_ON on firmware download failure"
This reverts commit 3ad833d71baac0d08946f6a795d91c4b29905b4d.

Reverting this commit to avoid the crash in case if the size of
the file read is zero.

Use case:
=========
When trying to create the calibration data files in the file system
by reading from the ART partition and if the partition has bad blocks,
then the file contents will be zero. This causes to hit the BUG_ON
while trying to read the files with zero size.

In this situation, we should do a graceful exit rather than BUG_ON.

Signed-off-by: Nitheesh Sekar <nsekar@codeaurora.org>
Change-Id: I0f21aee6a0f8db34d8eb73d92c698cee150e32c4