ipq40xx: clk: Fix for out-of-bound access at clk_notifier_register()
In clk_notifier_register(), if clk doesn't belongs to any clk_notifier in
global clk_notifier_list, cn will not point to a valid struct clk_notifier.
This patch fixes it by using a temp pointer to iterate and assign to cn, if
clk notifier found and then check the cn before dereferencing.

Signed-off-by: Karthick Shanmugham <kartshan@codeaurora.org>
Change-Id: I3edb1f044001fbbe9fbd237c14a4887cd9225818