net: cnss2: Unregister subsys notifier callback if subsystem_get fails
If subsystem_get fails in cnss_wlan_register_driver, subsys notifier
callback is not removed from the notifier chain. Since insmod of
wifi_3.0.ko will fail if subsystem_get fails, rmmod flow and
cnss_wlan_unregister_driver is not called and the notifier call chain is
never removed.

During the next insmod, the notifier callback is added to the existing
callback as a chain and next node will point to the same callback
function. This causes an infinite loop to happen and we get stuck with
continuous SUBSYS_BEFORE_POWERUP event from the SSR framework.

To fix this, unregister the notifier callback if subsytem_get fails in
cnss_wlan_register_driver.

Signed-off-by: Vignesh Viswanathan <viswanat@codeaurora.org>
Change-Id: I63d308702a115606f5f67d29e166dd524367a049