sound: ipq: use IRQ_DISABLE_LAZY flag for mbox irqs
This patch ensures that MBOX interrupts are using the
flag IRQ_DISABLE_LAZY.

Without this flag, the interrupt will be mark as disabled,
but the hardware will be left unmasked, during when any
pending interrupts will be delivered after the interrupt has
been marked disabled.

This is necessary because sometimes after mbox dma has been
stopped, the pending interrupts were delivered which caused
a kernel panic.

Signed-off-by: Selvam Sathappan Periakaruppan <speriaka@codeaurora.org>
Change-Id: Id72836deddb4b67f8a1071161dc24e50c698154d