drivers: qcom: wdt: Populate segment size dynamically
This patch updates the section size dynamically based on
number of sections for the mini dump utility.

When this is not done, there are certain sections for
which 1K memory is allocated, but size 2K is tried
during copy_to_user which should not be allowed.

Also, this patch fixes the for loop to iterate over
number of sections. Previously it was iterating from
0 till nsections which is not correct.

Signed-off-by: Selvam Sathappan Periakaruppan <speriaka@codeaurora.org>
Change-Id: I476dfcdf51b9f93794117b56f353c827167d1e0c