net: cnss2: Assert if subsystem_get fails
If subsystem_get fails, we will store the err value in
plat_priv->subsys_info.subsys_handle and return the err value to the
wifi driver. If the wifi driver tries to power on target again, it'll
fail because subsys_handle is not NULL.

To avoid this case and also to debug subsystem_get failures, trigger an
assert if subsystem_get fails.

Signed-off-by: Vignesh Viswanathan <viswanat@codeaurora.org>
Change-Id: I55f8658e5cddbf8d26a3857862d58a1342ddf4a6