mhi: flow control at mhi_netdev
Maple has only 2 ARM cores and 1 NSS core.
The core 0 does MHI netdet NAPI as well as NSS TX.
The core 1 does mhi-rmnet as well as NSS TX.
Whereas in csae of HK core 0 is used for mhi, 1 is
used for mhi-rmnet, core 2 and 3 are used for NSS TX.
And HK has 2 NSS cores. Both the ARM and NSS have
more capacity to consume more traffic. But in case of
Maple, due to reduced ARM and NSS Cores and also DDR being
only 512M, the cores are not having sufficient bandwidth
to consume higher througput, nor the memory has enough
capacity.

Hence dropping at MHI level if it crosses the threshold.

Change-Id: Ie2808627e5c8e24deac94db3bf6313e5d0c2407a
Signed-off-by: Manoharan Vijaya Raghavan <mraghava@codeaurora.org>