net: cnss2: Add support to download HDS & REGDB binaries
Add support to enable HDS and REGDB download prior to BDF download.
HDS binary file encompasses the sequence of API calls in HALPHY and also
the information about what are the APIs to be called for different events.
HDS bin is useful for debug where different sequences can be experimented
with just by replacing the modified bin instead of having to change the
code. REGDB binary consists of the regulatory database which earlier used
to be part of the BDF. Since it is something common across RDPs, it's
taken out of the BDF and formed as a separate binary. REGDB download
ideally would happen always but for now it is disabled by default using
module param infrastructure. HDS download would happen on demand where the
request can be enabled through cnsscli. In cases where cnss-daemon is
disabled, the request can be enabled through debugfs. Provision to enable
REGDB via cnsscli is also available.

Change-Id: If9ff2b9b48129560f5ad7fa57b71dfe9fa66b1bc
Signed-off-by: Sajini R <sajinir@codeaurora.org>